Do bootstrap calculations on each IPHC Series

To be called from within calc_iphc_ser_all(), once for each series.

boot_iphc(ser_year_rates, num.boots = 10000, seed_val = 42)

Arguments

ser_year_rates
Tibble of two columns containing just year and catch rates (either C_it or C_it20 depending on the Series).

num.boots
Number of bootstrap replicates to do.

seed_val
Seed for random number generator.

Value

Tibble containg one row for each year, with columns year, I_tBootMean, I_tBootLow, I_tBootHigh and I_tBootCV. If ser_year_rates is NULL or all counts are NA then return year = 2003 and all NA's (to not break later code).