Commercial CPUE index standardization

fit_cpue_index() fits a delta-Gamma commercial CPUE standardization model

predict_cpue_index() predicts from a delta-Gamma commercial CPUE standardization model

plot_cpue_index() plots a delta-Gamma commercial CPUE standardization model

tidy_cpue_index_coefs() extracts coefficients from a CPUE index standardization model

plot_cpue_index_coefs() plots coefficients from a CPUE index standardization model

plot_cpue_index_jk() "jackknifes" out terms one by one to test the sensitivity of the standardization model to any one term.

fit_cpue_index(dat, formula_binomial = cpue ~ year_factor,
  formula_gamma = spp_catch/hours_fished ~ year_factor)

predict_cpue_index(object, center = FALSE)

plot_cpue_index(predicted_dat, all_models = TRUE)

tidy_cpue_index_coefs(object, include_scale_pars = FALSE)

plot_cpue_index_coefs(object, regex = c(", base_[a-z]+", ""),
  type = c("delta", "tweedie"), ...)

plot_cpue_index_jk(object, terms = c("f(month)", "f(vessel)",
  "f(locality)", "f(depth)", "f(latitude)"), return_data = FALSE)

Arguments

dat
A data frame from tidy_cpue_index(), or a similarly formatted data frame.

formula_binomial
Formula for the binomial (presence-absence) model. See details.

formula_gamma
Formula for the Gamma (positive) model. See details.

object
Model output from fit_cpue_index().

center
Should the index be centered by subtracting the mean in link space?

predicted_dat
Input data frame, for example from predict_cpue_index().

all_models
Logical for whether to plot all models (TRUE) or just the combined model (FALSE).

include_scale_pars
Logical: include observation CV and random effect like standard deviations in parameter plot?

regex
A regular expression as a character vector of length 2 to transform the coefficient names from something (element 1) to something (element 2).

type
Delta or Tweedie?

...
Other arguments to pass to the tidy function.

terms
A character vector of terms to jackknife out.

return_data
Logical: should the data be returned?

Details

Formulas should be specified in the way a GLM formula would be specified in R or as a GLMM with up to 2 random intercepts would be specified in lme4::lmer(). E.g. pos_catch ~ year_factor + (1 | vessel) or cpue ~ year_factor + (1 | vessel) + (1 | locality). Your formula must have a predictor named year_factor for the other functions in this package to interact with the output properly. The fixed effect components of the binomial and Gamma components do not need to be the same, but the random effect components must be the same.

Because of how the formula parsing is performed, you cannot include any operations in the response variable. For example, you can't use catch/effort ~ year_factor. Instead you must create a new column, e.g. cpue and use that as the response: cpue ~ year_factor.

You may want to take advantage of the function gfplot::f() for conveniently turning predictors into factors with their base level set to the most common factor level and their factor names with consistent name lengths (e.g. '001', '020').

The model is internally fitted with TMB. As a result, you will have to have a C++ compiler installed to compile the models. On Windows, search for Rtools.

Note that for coefficients for predictors treated as factors, the coefficients represent the difference from the base level factor. For example, if January was the base month, month 02 represents the estimated difference between February and January. The default setup of the function gfplot::f() sets the reference level as the most common factor level.