Convert groundfish PBS data to a DLMtool data object

Takes the output from cache_pbs_data() and converts it into a DLMtool data object.

pbs2dlmtool_data(dat, name = "", area = "3[CD]+",
  survey = "SYN WCVI", max_year = 2017, min_mean_length = 10,
  length_bin_interval = 2)

Arguments

dat
A list object output from cache_pbs_data().

name
A name for the stock.

area
The groundfish and statistical area to subset the catch by.

survey
A survey abbreviation to include as the relative index of abundance.

max_year
The most recent year of data to include.

min_mean_length
The minimum number of samples to include a mean length measurement for a given year.

length_bin_interval
An interval for the length bins.

Value

An S4 object of class DLMtool Data.

Examples

# NOT RUN {
cache_pbs_data("pacific ocean perch", unsorted_only = FALSE)
d <- readRDS("pacific-ocean-perch.rds")
pbs2dlmtool_data(d)
# }