Plot survey relative biomass mass index

Plot survey relative biomass mass index

tidy_survey_index(dat, survey = c("SYN WCHG", "SYN HS", "SYN QCS",
  "SYN WCVI", "HBLL OUT N", "HBLL OUT S", "HBLL INS N", "HBLL INS S",
  "MSA HS", "IPHC FISS"), min_years = 3, year_range = NULL)

plot_survey_index(dat, col = brewer.pal(9, "Greys")[c(3, 7)],
  max_cv = 0.4, max_set_fraction = 0.05, xlim = NULL,
  survey_cols = NULL, scale = TRUE, year_increment = 5,
  hide_y_axis = FALSE)

Arguments

dat
For tidy_survey_index(), data from get_survey_index(). For plot_survey_index(), data from tidy_survey_index().

survey
A character vector of survey abbreviations to include.

min_years
The minimum number of years for a survey before it is included.

year_range
If not NULL, a numeric vector of length 2 giving the range of years to plot.

col
A vector of two colours for the lines and shading.

max_cv
A coefficient of variation above which a panel will be shaded as more uncertain.

max_set_fraction
A fraction of positive sets above which a panel will be shaded as more uncertain.

xlim
If not NULL, the x axis limits.

survey_cols
If not NULL, a named character vector of colors for the various surveys.

scale
Logical: scale the biomass by the maximum?

year_increment
Increment for the year x axis.

hide_y_axis
Logical: hide the y axis ticks and labels?

Examples

# NOT RUN {
get_survey_index("lingcod") %>%
  tidy_survey_index() %>%
  plot_survey_index()
# }