Tidy commercial PBS CPUE data

This function determines the qualifying "fleet" for CPUE analyses.

tidy_cpue_index(dat, species_common, year_range = c(1996,
  as.numeric(format(Sys.Date(), "%Y")) - 1), lat_range = c(48, Inf),
  min_positive_tows = 100, min_positive_trips = 4, min_yrs_with_trips = 4,
  area_grep_pattern = "5[CDE]+", lat_band_width = 0.2,
  depth_band_width = 50, clean_bins = TRUE, depth_bin_quantiles = c(0.01,
  0.99), lat_bin_quantiles = c(0.01, 0.99), gear = "BOTTOM TRAWL")

Arguments

dat
An input data frame from get_cpue_index().

species_common
The species common name.

year_range
The range of years to include.

lat_range
The range of latitudes to include.

min_positive_tows
The minimum number of positive tows over all years.

min_positive_trips
The minimum number of annual positive trips.

min_yrs_with_trips
The number of years in which the min_positive_trips criteria needs to be met.

area_grep_pattern
A regular expression to extract the management areas of interest.

lat_band_width
The latitude bandwidths in degrees.

depth_band_width
The depth band widths in m.

clean_bins
Logical. Should the depth and latitude bands be rounded to the nearest clearn value as defined by lat_band_width or depth_band_width? Internally, these use, for example: gfplot:::round_down_even(lat_range[1], lat_band_width).

depth_bin_quantiles
Quantiles for the depth bands. Values above and below these quantiles will be discarded.

lat_bin_quantiles
Quantiles for the latitude bands. Values above and below these quantiles will be discarded.

gear
Gear types

Examples

# NOT RUN {
d <- get_cpue_index(gear = "bottom trawl")
walleye <- tidy_cpue_index(d, "walleye pollock",
  area_grep_pattern = "5[CDE]+")
# }